requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
cv2-extras==0.5.4
cycler==0.10.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.52
peewee==3.14.4
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.5
